Add circle shape action
