+++ lib/app_constants.dart
@@ -2,5 +2,6 @@ import 'package:flutter/material.dart';
 
 class AppConstants {
   static const String appName = 'Touch Figure';
+  static const Color circleColor = Colors.blue;
   static const Color seedColor = Colors.deepPurple;
 }
